+++ tts_engine.py
@@ -47,7 +47,7 @@ except ImportError:
     PYDUB_AVAILABLE = False
     AudioSegment = None
 
-VERSION = "5.2.1"
+VERSION = "6.0.0-dev"
 
 # 当前平台
 _PLATFORM = platform.system()  # "Darwin" / "Windows" / "Linux"